% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concalculations.R
\name{beerlambert}
\alias{beerlambert}
\title{beerlambert}
\usage{
beerlambert(A, epsilon, l)
}
\arguments{
\item{A}{absorptie}

\item{epsilon}{molaire absorptiviteitscoëfficiënt}

\item{l}{padlengte van het monster}
}
\value{
Numeric
}
\description{
Berekent de concentratie uit de Beer-Lambert wet
}
\examples{
beerlambert(A = 1, epsilon = 0.5, l = 1)
}
