% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concalculations.R
\name{massaprocent}
\alias{massaprocent}
\title{Massaprocent}
\usage{
massaprocent(x, y)
}
\arguments{
\item{x}{massa van de opgeloste stof}

\item{y}{massa van de oplossing}
}
\value{
Numeric
}
\description{
Berekent het massaprocent
}
\examples{
massaprocent(5, 100)
}
