% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concalculations.R
\name{molariteit}
\alias{molariteit}
\title{Molariteit}
\usage{
molariteit(n, V)
}
\arguments{
\item{n}{aantal mol opgeloste stof}

\item{V}{volume van de oplossing in L}
}
\value{
Numeric
}
\description{
Berekent de molariteit
}
\examples{
molariteit(n = 2, V = 1)
}
