% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conculations.R
\name{verdunningsfactor}
\alias{verdunningsfactor}
\title{Verdunningsfactor}
\usage{
verdunningsfactor(x, y)
}
\arguments{
\item{x}{begin concentratie}

\item{y}{end concentratie}
}
\value{
Numeric
}
\description{
Berekent de verdunningsfactor
}
\examples{
x <- 10
y <- 15
verdunningsfactor(x, y)
}
